[DEBUG] libulog: prevent __ulog_default_cookie to be added twice in list.
If no ULOG_TAG defined, ulog use the __ulog_default_cookie.
This cookie is initialized with a log level = -1 at startup (ie not initialized)
and it is already registered in the list.
At it first use, ulog_init_cookie add it in the list whereas it was already added.
This patch remove the startup add of the default cookie in the list.