[DEV] libulog: add ULOGx_THROTTLE macros to prevent log flooding
These macros are similar to ULOGx macros except that the first
argument is a delay expressed in milliseconds. When a message is
logged, subsequent invocations of the same logging macro do not
produce messages until the specified delay has expired.
This is useful to prevent log buffer flooding.
When messages have been masked due to throttling, a counter is
prepended to the next printed message to indicate the number of
masked logs.