[DEV] libulog: add ULOGx_CHANGE macros for logging value transitions
These macros are similar to ULOGx macros except that the first
argument is a scalar value. Invoking these macros only produce
messages when the value has changed, or when invoked for the first
time. This is useful for logging only the transitions of a value.
Note: the value is cast to type uintptr_t for change comparison.